\name{Function2 of Exercise3}
\alias{regression}

\title{Regression Function}
\description{
Calculate the projection of Co2, methane and nitrous oxide by country. Use a linear model to this computation (lm): For this, decade 1990-2000 has the value 1, 2000-2010 the value of 2 and so on, consecutively it can be transformed into a regression model. Moreover in the plot it is indicated what decades or countries are represented.
}
\usage{
regression(fcountry)
}

\author{
Heribert Roig, Antoni Bosch, Nuria Bosch.
}

\examples{
regression("China") }
