\name{Function 3 of Exericse3}
\alias{statistics}

\title{Statistics Function}
\description{
It projects the future of Co2, methane and nitrous oxide and calculate statistics by country giving the following information: Will Increment their Co2?, Expected increase in million tonnes, Expected percentage of the increase and Standard deviation and R-squared value of the regression model.
}
\usage{
statistics(fcountry)
}

\author{
Heribert Roig, Antoni Bosch, Nuria Bosch.
}

\examples{
statistics(c("Spain", "China", "France", "Andorra")) }
